Also run Poetry-based workflow when triggered from push or PR
This only occurs in the case of push and PR within the DevOps repo itself.
But it's still necessary to cover this case, and it should use the Poetry-
based variant, as this is based on the DevOps branch that already uses
Poetry. Maybe a cleaner solution could be done with ENV variables? IDK