Default to use poetry on push and PR (#51)
* Default to use poetry on push and PR

This is necessary because input parameters only exist for workflow_call
and workflow_dispatch, so it's not really possible to set a default
value when the action is triggered by a push or pull request. However,
since the sample project also uses poetry, those triggers should use it as
well. This should solve that.

See PR #51 for detailed explanation.

Also contains:

* Fix workflow names

* Rollout and cleanup