Try setting up poetry manually
Looking at the previously used action workflow's source code
(https://github.com/abatilo/actions-poetry/blob/master/action.yml),
this seems to be all it does anyway. So probably better to just do it
manually and have full control, if that's all we need.

This will most likely not solve the macOS issue, but at least it should
make it easier to debug and fix it...