Remove non-Poetry options from workflows
This is now possible because all repos that use these workflows use them
with Poetry anyway, and we're not planning to add new ones that don't.

Still, I'm keeping the input parameter around so that things don't break.
We can slowly start removing the `with: poetry: true` options in all the
repos that use these and then fully remove it here. For the
`workflow_dispatch` (aka manual trigger) case, it can be removed already.